requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib
python-dotenv
bcrypt
alembic
psycopg2-binary  # PostgreSQLを使う場合
